Short fixes

+++ src/Maths/Silk.NET.Maths/Scalar.BaseOps.cs
@@ -1461,7 +1461,7 @@ namespace Silk.NET.Maths
             {
                 if (typeof(T) == typeof(short))
                 {
-                    return (T) (object) ((short) (object) left + (short) (object) right);
+                    return (T) (object) (short) ((short) (object) left + (short) (object) right);
                 }
 
                 return UShort(left, right);
@@ -1472,7 +1472,7 @@ namespace Silk.NET.Maths
             {
                 if (typeof(T) == typeof(ushort))
                 {
-                    return (T) (object) ((ushort) (object) left + (ushort) (object) right);
+                    return (T) (object) (ushort) ((ushort) (object) left + (ushort) (object) right);
                 }
 
                 return Int(left, right);
@@ -1601,7 +1601,7 @@ namespace Silk.NET.Maths
             {
                 if (typeof(T) == typeof(short))
                 {
-                    return (T) (object) ((short) (object) left - (short) (object) right);
+                    return (T) (object) (short) ((short) (object) left - (short) (object) right);
                 }
 
                 return UShort(left, right);
@@ -1612,7 +1612,7 @@ namespace Silk.NET.Maths
             {
                 if (typeof(T) == typeof(ushort))
                 {
-                    return (T) (object) ((ushort) (object) left - (ushort) (object) right);
+                    return (T) (object) (ushort) ((ushort) (object) left - (ushort) (object) right);
                 }
 
                 return Int(left, right);
@@ -1741,7 +1741,7 @@ namespace Silk.NET.Maths
             {
                 if (typeof(T) == typeof(short))
                 {
-                    return (T) (object) ((short) (object) left * (short) (object) right);
+                    return (T) (object) (short) ((short) (object) left * (short) (object) right);
                 }
 
                 return UShort(left, right);
@@ -1752,7 +1752,7 @@ namespace Silk.NET.Maths
             {
                 if (typeof(T) == typeof(ushort))
                 {
-                    return (T) (object) ((ushort) (object) left * (ushort) (object) right);
+                    return (T) (object) (ushort) ((ushort) (object) left * (ushort) (object) right);
                 }
 
                 return Int(left, right);
@@ -1881,7 +1881,7 @@ namespace Silk.NET.Maths
             {
                 if (typeof(T) == typeof(short))
                 {
-                    return (T) (object) ((short) (object) left / (short) (object) right);
+                    return (T) (object) (short) ((short) (object) left / (short) (object) right);
                 }
 
                 return UShort(left, right);
@@ -1892,7 +1892,7 @@ namespace Silk.NET.Maths
             {
                 if (typeof(T) == typeof(ushort))
                 {
-                    return (T) (object) ((ushort) (object) left / (ushort) (object) right);
+                    return (T) (object) (ushort) ((ushort) (object) left / (ushort) (object) right);
                 }
 
                 return Int(left, right);
