+++ src/Maths/Silk.NET.Maths/Scalar.Inverse.cs
@@ -86,7 +86,7 @@ namespace Silk.NET.Maths
             {
                 if (typeof(T) == typeof(ushort))
                 {
-                    return (T) (object) (((ushort) 1) / (ushort) (object) x);
+                    return (T) (object) (ushort) (((ushort) 1) / (ushort) (object) x);
                 }
 
                 return Short(x);
@@ -97,7 +97,7 @@ namespace Silk.NET.Maths
             {
                 if (typeof(T) == typeof(short))
                 {
-                    return (T) (object) (((short) 1) / (short) (object) x);
+                    return (T) (object) (short) (((short) 1) / (short) (object) x);
                 }
 
                 return UInt(x);
